marketplace/broker/requirements.txt
Flask>=0.10.1
Flask-PyMongo>=0.3.1
PyJWT>=1.3.0
requests>=2.7.0